[Refactor] 병합

+++ server/src/main/java/com/codinghaezo/stackOverFlow/domain/question/service/QuestionService.java
@@ -1,14 +1,19 @@
 package com.codinghaezo.stackOverFlow.domain.question.service;
 
+import com.codinghaezo.stackOverFlow.Utils.UriCreator;
 import com.codinghaezo.stackOverFlow.domain.question.entity.Question;
 import com.codinghaezo.stackOverFlow.domain.question.repository.QuestionRepository;
+import com.codinghaezo.stackOverFlow.exception.BusinessLogicException;
+import com.codinghaezo.stackOverFlow.exception.ExceptionCode;
+import com.codinghaezo.stackOverFlow.member.Member;
+import com.codinghaezo.stackOverFlow.member.MemberRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
+import java.net.URI;
 
 @Service
 @Transactional
@@ -16,17 +21,27 @@ public class QuestionService {
 
     private final QuestionRepository questionRepository;
 
-    public QuestionService(QuestionRepository questionRepository) {
+
+    private final MemberRepository memberRepository;
+
+    public QuestionService(QuestionRepository questionRepository, MemberRepository memberRepository) {
+
         this.questionRepository = questionRepository;
+        this.memberRepository = memberRepository;
     }
 
-    public Question createQuestion(Question question) {
-        return questionRepository.save(question);
+    public URI createQuestion(String email, Question question) {
+        Member author = findMemberByEmail(email);
+        question.setAuthor(author);
+        long questionId = questionRepository.save(question).getId();
+        return UriCreator.createUri("/questions", questionId);
     }
 
     public Question findQuestion(long questionId) {
-        Optional<Question> foundQuestion = questionRepository.findById(questionId);
-        return foundQuestion.orElseThrow(RuntimeException::new);
+        Question foundQuestion = questionRepository.findById(questionId)
+            .orElseThrow(() -> new BusinessLogicException(ExceptionCode.QUESTION_NOT_FOUND));
+        foundQuestion.setViews(questionRepository.updateView(questionId));
+        return foundQuestion;
     }
 
     public Page<Question> findQuestions(int page, int size) {
@@ -34,25 +49,41 @@ public class QuestionService {
         return questionRepository.findAll(pageRequest);
     }
 
-    public Question updateQuestion(Question question) {
-        Question foundQuestion = findQuestion(question.getId());
-        String title = question.getTitle();
-        if (title != null) {
-            foundQuestion.setTitle(title);
-        }
-        String bodyProblem = question.getBodyProblem();
-        if (bodyProblem != null) {
-            foundQuestion.setBodyProblem(bodyProblem);
-        }
-        String bodyExpecting = question.getBodyExpecting();
-        if (bodyExpecting != null) {
-            foundQuestion.setBodyExpecting(bodyExpecting);
-        }
-        return questionRepository.save(foundQuestion);
+    public Question updateQuestion(long questionId, Question question, String principalEmail) {
+        Question foundQuestion = findQuestion(questionId);
+        verifyAuthor(principalEmail, foundQuestion);
+        Question updatedQuestion = Question.builder()
+            .id(questionId)
+            .title((question.getTitle() == null)
+                ? foundQuestion.getTitle()
+                : question.getTitle())
+            .bodyProblem((question.getBodyProblem() == null)
+                ? foundQuestion.getBodyProblem()
+                : question.getBodyProblem())
+            .bodyExpecting((question.getBodyExpecting() == null)
+                ? foundQuestion.getBodyExpecting()
+                : question.getBodyExpecting())
+            .build();
+        return questionRepository.save(updatedQuestion);
     }
 
-    public void deleteQuestion(long questionId) {
+    public void deleteQuestion(long questionId, String principalEmail) {
         Question foundQuestion = findQuestion(questionId);
+        verifyAuthor(principalEmail, foundQuestion);
         questionRepository.delete(foundQuestion);
     }
+
+
+    private Member findMemberByEmail(String email) {
+        return memberRepository.findByEmail(email)
+            .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+    }
+
+    private void verifyAuthor(String principalEmail, Question question) {
+        String authorEmail = question.getAuthor().getEmail();
+        if (!principalEmail.equals(authorEmail)) {
+            throw new BusinessLogicException(ExceptionCode.CANNOT_CHANGE_QUESTION);
+        }
+    }
+
 }
